Rename match-to-case to naive-match-to-case and add match-to-case stub
